+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Allow configuring labels and annotations for JMX authentication secrets
 * Enable Cruise Control anomaly.detection configurations
 * Add support for building connector images from the Maven coordinates
+* Allow Kafka Connect Build artifacts to be downloaded from insecure servers (#5542)
 
 ### Changes, deprecations and removals